Day14/UnitConverter.java
public class UnitConverter {


    public double centimeterToCentimeter(double magnitudeInCentimeter) {
        return magnitudeInCentimeter;
    }

    public double meterToCentimeter(double magnitudeInMeter) {
        return magnitudeInMeter * 100;
    }

    public double kilometerToCentimeter(double magnitudeInKilometer) {
        return magnitudeInKilometer * 100000;
    }

    public double centimeterToKilometer(double magnitudeInCentimeter) {
        return magnitudeInCentimeter / 100000;
    }

    public double centimeterToMeter(double magnitudeInCentimeter) {
        return magnitudeInCentimeter / 100;
    }
}
